lotto_dashboard/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.3.0
requests==2.32.3
requests-cache==1.2.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
gunicorn==21.2.0
